Make Order Pitch Screen

+++ iPitch/iPitch/Services/PitchService.swift
@@ -13,6 +13,10 @@ import FirebaseDatabase
 
 class PitchService: NSObject {
     
+    enum PitchServiceError: Error {
+        case userNotFound
+    }
+    
     static let shared = PitchService()
     private let ref = FIRDatabase.database().reference().child("pitches")
     fileprivate let locationManager = CLLocationManager()
@@ -27,13 +31,14 @@ class PitchService: NSObject {
         }
     }
     
-    func getPitch(radius: Double?, districtId: Int?, timeFrom: Date?,
+    func getPitch(searchText: String?, radius: Double?, districtId: Int?, timeFrom: Date?,
         timeTo: Date?, completion: @escaping ([Pitch]) -> Void) {
         getAllPitches { (pitches) in
             let pitchesFilter = pitches.filter({
-                [weak self] (pitch) -> Bool in
-                if let radius = radius,
-                    let location = self?.locationManager.location {
+                [unowned self] (pitch) -> Bool in
+                if var radius = radius,
+                    let location = self.locationManager.location {
+                    radius = radius * 1000
                     guard location.distance(
                         from: CLLocation(latitude: pitch.latitude,
                         longitude: pitch.longitude)) <= radius else {
@@ -55,10 +60,32 @@ class PitchService: NSObject {
                         return false
                     }
                 }
+                if var searchText = searchText {
+                    searchText = searchText.uppercased()
+                    guard pitch.name.uppercased().contains(searchText) ||
+                        pitch.address.uppercased().contains(searchText) ||
+                        pitch.phone.uppercased().contains(searchText) else {
+                        return false
+                    }
+                }
                 return true
             })
-            OperationQueue.main.addOperation {
-                completion(pitchesFilter)
+            guard let location = self.locationManager.location else {
+                DispatchQueue.main.async {
+                    completion(pitchesFilter)
+                }
+                return
+            }
+            let pitchesSorted = pitchesFilter.sorted(by: {
+                let location0 = CLLocation(latitude: $0.latitude,
+                    longitude: $0.longitude)
+                let location1 = CLLocation(latitude: $1.latitude,
+                    longitude: $1.longitude)
+                return location.distance(from: location0) <
+                    location.distance(from: location1)
+            })
+            DispatchQueue.main.async {
+                completion(pitchesSorted)
             }
         }
     }
@@ -84,19 +111,66 @@ class PitchService: NSObject {
 
     func create(pitch: Pitch, photo: UIImage?,
         completion: @escaping (Error?) -> Void) {
-        if let userId = FIRAuth.auth()?.currentUser?.uid {
+        guard let userId = FIRAuth.auth()?.currentUser?.uid else {
+            DispatchQueue.main.async {
+                completion(PitchServiceError.userNotFound)
+            }
+            return
+        }
+        var json = pitch.toJSON()
+        if json["id"] != nil {
+            json.removeValue(forKey: "id")
+        }
+        if json["photoPath"] != nil {
+            json.removeValue(forKey: "photoPath")
+        }
+        json["ownerId"] = userId
+        ref.child(userId).childByAutoId().setValue(json,
+            withCompletionBlock: { (error, ref) in
+            if error == nil, let photo = photo {
+                let photoPath = "images/pitches/\(ref.key).jpg"
+                StorageService.shared.uploadImage(image: photo,
+                    path: photoPath, completion: {
+                    [weak self] (error, url) in
+                    if let url = url {
+                        self?.ref.child("\(ref.key)/photoPath").setValue(
+                            url.absoluteString, withCompletionBlock: {
+                            (error, ref) in
+                            DispatchQueue.main.async {
+                                completion(error)
+                            }
+                        })
+                    } else {
+                        DispatchQueue.main.async {
+                            completion(error)
+                        }
+                    }
+                })
+            } else {
+                DispatchQueue.main.async {
+                    completion(error)
+                }
+            }
+        })
+    }
+    
+    func update(pitch: Pitch, photo: UIImage?,
+        completion: @escaping (Error?) -> Void) {
+        guard let userId = FIRAuth.auth()?.currentUser?.uid else {
+            DispatchQueue.main.async {
+                completion(PitchServiceError.userNotFound)
+            }
+            return
+        }
+        if let pitchId = pitch.id {
             var json = pitch.toJSON()
             if json["id"] != nil {
                 json.removeValue(forKey: "id")
             }
-            if json["photoPath"] != nil {
-                json.removeValue(forKey: "photoPath")
-            }
-            json["ownerId"] = userId
-            ref.childByAutoId().setValue(json, withCompletionBlock: {
-                (error, ref) in
+            ref.child("\(userId)/\(pitchId)").setValue(json,
+                withCompletionBlock: { (error, ref) in
                 if error == nil, let photo = photo {
-                    let photoPath = "images/pitches/\(ref.key).jpg"
+                    let photoPath = "images/pitches/\(pitchId).jpg"
                     StorageService.shared.uploadImage(image: photo,
                         path: photoPath, completion: {
                         [weak self] (error, url) in
@@ -104,44 +178,18 @@ class PitchService: NSObject {
                             self?.ref.child("\(ref.key)/photoPath").setValue(
                                 url.absoluteString, withCompletionBlock: {
                                 (error, ref) in
-                                OperationQueue.main.addOperation {
+                                DispatchQueue.main.async {
                                     completion(error)
                                 }
                             })
                         } else {
-                            OperationQueue.main.addOperation {
+                            DispatchQueue.main.async {
                                 completion(error)
                             }
                         }
                     })
                 } else {
-                    OperationQueue.main.addOperation {
-                        completion(error)
-                    }
-                }
-            })
-        }
-    }
-    
-    func update(pitch: Pitch, photo: UIImage?,
-        completion: @escaping (Error?) -> Void) {
-        if let pitchId = pitch.id {
-            var json = pitch.toJSON()
-            if json["id"] != nil {
-                json.removeValue(forKey: "id")
-            }
-            ref.child(pitchId).setValue(json, withCompletionBlock: {
-                (error, ref) in
-                if error == nil, let photo = photo {
-                    let photoPath = "images/pitches/\(pitchId).jpg"
-                    StorageService.shared.uploadImage(image: photo,
-                        path: photoPath, completion: { (error, url) in
-                        OperationQueue.main.addOperation {
-                            completion(error)
-                        }
-                    })
-                } else {
-                    OperationQueue.main.addOperation {
+                    DispatchQueue.main.async {
                         completion(error)
                     }
                 }
@@ -150,30 +198,40 @@ class PitchService: NSObject {
     }
     
     func delete(pitch: Pitch, completion: @escaping (Error?) -> Void) {
+        guard let userId = FIRAuth.auth()?.currentUser?.uid else {
+            DispatchQueue.main.async {
+                completion(PitchServiceError.userNotFound)
+            }
+            return
+        }
         if let pitchId = pitch.id {
-            ref.child(pitchId).removeValue(completionBlock: { (error, ref) in
-                OperationQueue.main.addOperation {
+            ref.child("\(userId)/\(pitchId)").removeValue(completionBlock: {
+                (error, ref) in
+                DispatchQueue.main.async {
                     completion(error)
                 }
             })
         }
         if let photoPath = pitch.photoPath {
             StorageService.shared.deleteImage(path: photoPath,
-                completion: { (error) in
-                //
-            })
+                completion:nil)
         }
     }
 
     private func getAllPitches(completion: @escaping ([Pitch]) -> Void) {
         ref.observeSingleEvent(of: .value, with: { (snapshot) in
             var pitches = [Pitch]()
-            if let pitchesJSON = snapshot.value as? [String: Any] {
-                for (key, value) in pitchesJSON {
-                    if var pitchJSON = value as? [String: Any] {
-                        pitchJSON["id"] = key
-                        if let pitch = Pitch(JSON: pitchJSON) {
-                            pitches.append(pitch)
+            if let usersJSON = snapshot.value as? [String: Any] {
+                for (_, value) in usersJSON {
+                    guard let pitchesJSON = value as? [String: Any] else {
+                        continue
+                    }
+                    for (key, value) in pitchesJSON {
+                        if var pitchJSON = value as? [String: Any] {
+                            pitchJSON["id"] = key
+                            if let pitch = Pitch(JSON: pitchJSON) {
+                                pitches.append(pitch)
+                            }
                         }
                     }
                 }
